Create new $wp_roles object in install_blog() to avoid protected _init() call
Merges [30242] to the 4.0 branch.

Our call to `$wp_roles->_init()` relied on the `__call()` method in `WP_Roles` to handle the link to the protected method. This works back to PHP 5.2.9, when a bug was fixed allowing access to protected methods through this exact approach.

`install_blog()` needs a fresh `$wp_roles` object after `populate_roles()` resets everything in its path. We can create this new object from scratch, effectively doing the same thing with the call to `_init()` via the constructor.

props jeremyfelt.
fixes #29692.


git-svn-id: https://develop.svn.wordpress.org/branches/4.0@30255 602fd350-edb4-49c9-b593-d223f7449a82