Meta boxes: Don't assume that callback args are an array.
While the documentation for `add_meta_box()` specifices that `$callback_args` should be an array, this has never been enforced, and we have workarounds in place for when it's passed as something other than an array.

Rather than break sites that are passing unexpected data, we can quietly just allow for it, instead.

Props johnjamesjacoby, birgire.
Fixes #45206.



git-svn-id: https://develop.svn.wordpress.org/branches/5.0@43838 602fd350-edb4-49c9-b593-d223f7449a82