Twenty Thirteen: Fix button block custom colors.
Simplify some of the button block’s CSS selectors in the theme, to make sure the default colors don’t override the custom colors on the front end.

Also make sure the button’s gradient is not applied when a custom background color is added, since it will hide it. 

Fixes #45431.



git-svn-id: https://develop.svn.wordpress.org/branches/5.0@44195 602fd350-edb4-49c9-b593-d223f7449a82