Editor: Allow the “Add Media” button tooltip to be translated in the Classic block.
In the Classic Editor, the “Add Media” button was outside of the TinyMCE editor. In the Block Editor, the “Add Media” button was moved into the TinyMCE toolbar in the Classic block, but the tooltip was not added to the list of translatable strings. This adds “Add Media” to that list.

The corresponding keyboard shortcut is also specified for the “Add Media” button.

Props afercia.

Fixes #45788.

git-svn-id: https://develop.svn.wordpress.org/trunk@44398 602fd350-edb4-49c9-b593-d223f7449a82