Build/Test Tools: Explicitly specify a version number in the `.nvmrc` file for the 5.0 branch.
This restores the ability to run NodeJS related tasks when using `nvm install` or `nvm use`.

The alias `lts/*` currently resolves to NodeJS 12.x (and will continue to change as newer versions are released). The 10.x version of NodeJS is the highest version supported in the 5.0 branch.

This also removes the explicit version when running `nvm install` during automated testing. The command will now fall back to the version in the `.nvmrc` file.

See #51603.

git-svn-id: https://develop.svn.wordpress.org/branches/5.0@49277 602fd350-edb4-49c9-b593-d223f7449a82