Twenty Twenty-One: Re-add `px` unit to the admin bar height custom property.
The `px` unit added in [50388] to prevent invalid values when using `calc` functions was inadvertently removed in [50972].

This re-adds the unit to ensure the issue reported in #52564 remains fixed. The same stylelint rule adjustment made in [50388] to `.stylelintrc-css.json` has also been copied over to `stylelintrc.json` file,  which is the configuration file used when linting `.scss` files.

Props ocean90.
See #52624, #52564.

git-svn-id: https://develop.svn.wordpress.org/trunk@50975 602fd350-edb4-49c9-b593-d223f7449a82