External Libraries: Update the phpass library to version `0.5`.
This updates the phpass library from version `0.3` to the latest version of `0.5`. This is a minor update that includes some adjustments to be more friendly with newer versions of PHP.

The changes made in [30466] have also been applied to the new version of the library. This should now be the only WordPress specific modification do this library.

Props ayeshrajans, dd32, otto42, desrosj.
Fixes #51549.

git-svn-id: https://develop.svn.wordpress.org/trunk@51008 602fd350-edb4-49c9-b593-d223f7449a82